pyproject.toml
[project]
name = "goal-watcher"
version = "0.1.0"
description = "SmartThings app that triggers home events when a Scottish football team scores a goal"
requires-python = ">=3.14"
dependencies = [
    "aws-cdk-lib>=2.180.0",
    "aws-lambda-powertools>=3.7.0",
    "boto3>=1.42.77",
    "cdk-nag>=2.35.0",
    "constructs>=10.4.0",
    "httpx>=0.28.0",
    "pydantic>=2.11.0",
]

[dependency-groups]
dev = [
    "faker>=40.11.1",
    "mypy>=1.15.0",
    "polyfactory>=2.19.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.1.0",
    "pytest-socket>=0.7.0",
    "ruff>=0.11.0",
    "types-boto3[dynamodb,events,lambda]>=1.36.0",
]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = [
    "A", "B", "C", "C4", "DTZ", "E", "F", "FURB", "G", "I", "LOG", "N",
    "PERF", "PL", "PT", "PTH", "Q", "RET", "RSE", "RUF", "S", "T", "TID",
    "TRY", "UP", "W",
]
ignore = ["E501", "TRY003"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["PLR0913", "PLR2004", "PT012", "PT019", "S101", "S106", "S107"]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--disable-socket --allow-unix-socket --cov=app --cov-branch --cov-report=term-missing --cov-fail-under=60"
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
